treewide: fix compiler format specifiers, typecheck warnings, and Makefile duplication

+++ net/qrtr/af_qrtr.c
@@ -812,8 +812,8 @@ static int qrtr_node_enqueue(struct qrtr_node *node, struct sk_buff *skb,
 		rc = skb_put_padto(skb, ALIGN(len, 4) + sizeof(*hdr));
 
 	if (rc) {
-		pr_err("%s: failed to pad size %lu to %lu rc:%d\n", __func__,
-		       skb->len, ALIGN(skb->len, 4), rc);
+		pr_err("%s: failed to pad size %u to %u rc:%d\n", __func__,
+		       skb->len, (unsigned int)ALIGN(skb->len, 4), rc);
 	}
 
 	if (!rc) {
